requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
defusedxml==0.7.1
distro==1.9.0
fastapi==0.115.13
flask==3.1.1
gitdb==4.0.12
gitpython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jira==3.8.0
jiter==0.10.0
markupsafe==3.0.2
oauthlib==3.3.1
ollama==0.5.1
openai==1.91.0
packaging==25.0
pillow==11.2.1
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.34.3
werkzeug==3.1.3
